renamed resource interface, moved to DynamoViewModel
