--- src/DynamoCore/Interfaces/IHostApplication.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Dynamo.Interfaces
-{
-    public enum ResourceName
-    {
-        StartPageLogo,
-        AboutBoxLogo
-    }
-
-    public interface IHostApplication
-    {
-        string GetImageSource(ResourceName resourceName);
-        string GetStringResource(ResourceName resourceName);
-    }
-}
