+++ src/DynamoCoreWpf/Interfaces/IBrandingResourceProvider.cs
@@ -1,0 +1,16 @@
+﻿using System.Windows.Media;
+
+namespace Dynamo.Wpf.Interfaces
+{
+    public enum ResourceName
+    {
+        StartPageLogo,
+        AboutBoxLogo
+    }
+
+    public interface IBrandingResourceProvider
+    {
+        ImageSource GetImageSource(ResourceName resourceName);
+        string GetStringResource(ResourceName resourceName);
+    }
+}
