+++ src/DynamoCoreWpf/ViewModels/Core/DynamoViewModel.cs
@@ -19,6 +19,7 @@ using Dynamo.Selection;
 using Dynamo.Services;
 using Dynamo.UpdateManager;
 using Dynamo.Utilities;
+using Dynamo.Wpf.Interfaces;
 using Dynamo.Wpf.UI;
 using DynamoUnits;
 
@@ -26,6 +27,7 @@ using DynCmd = Dynamo.ViewModels.DynamoViewModel;
 using System.Reflection;
 using Dynamo.Wpf.Properties;
 using DynamoUtilities;
+using ResourceName = Dynamo.Wpf.Interfaces.ResourceName;
 
 namespace Dynamo.ViewModels
 {
@@ -425,6 +427,12 @@ namespace Dynamo.ViewModels
             public IVisualizationManager VisualizationManager { get; set; }
             public IWatchHandler WatchHandler { get; set; }
             public DynamoModel DynamoModel { get; set; }
+
+            /// <summary>
+            /// This property is initialized if there is an external host application
+            /// at startup in order to be used to pass in host specific resources to DynamoModel
+            /// </summary>
+            public IBrandingResourceProvider BrandingResourceProvider { get; set; }
         }
 
         public static DynamoViewModel Start()
@@ -438,12 +446,13 @@ namespace Dynamo.ViewModels
             var vizManager = startConfiguration.VisualizationManager ?? new VisualizationManager(model);
             var watchHandler = startConfiguration.WatchHandler ?? new DefaultWatchHandler(vizManager, 
                 model.PreferenceSettings);
-            
-            return new DynamoViewModel(model, watchHandler, vizManager, startConfiguration.CommandFilePath);
+
+            return new DynamoViewModel(model, watchHandler, vizManager, startConfiguration.CommandFilePath,
+                startConfiguration.BrandingResourceProvider);
         }
 
         protected DynamoViewModel(DynamoModel dynamoModel, IWatchHandler watchHandler,
-            IVisualizationManager vizManager, string commandFilePath)
+            IVisualizationManager vizManager, string commandFilePath, IBrandingResourceProvider resourceProvider)
         {
             // initialize core data structures
             this.model = dynamoModel;
@@ -458,6 +467,8 @@ namespace Dynamo.ViewModels
 
             // Start page should not show up during test mode.
             this.ShowStartPage = !DynamoModel.IsTestMode;
+
+            this.BrandingResourceProvider = resourceProvider ?? new DefaultBrandingResourceProvider();
 
             //add the initial workspace and register for future 
             //updates to the workspaces collection
@@ -2144,23 +2155,16 @@ namespace Dynamo.ViewModels
 
         public DynamoViewModel ViewModel { get { return this; } }
 
+        public IBrandingResourceProvider BrandingResourceProvider { get; private set; }
+
         public ImageSource AboutBoxIcon
         {
             get
             {
-                if (model.HostApplication != null)
-                    return new BitmapImage(
-                        new Uri(model.HostApplication.GetImageSource(ResourceName.AboutBoxLogo)));
-                
-                return new BitmapImage(
-                    new Uri(@"pack://application:,,,/DynamoCoreWpf;component/UI/Images/AboutWindow/logo_about.png",
-                        UriKind.Absolute));
+                return BrandingResourceProvider.GetImageSource(ResourceName.AboutBoxLogo);
             }
         }
 
         #endregion
-
-
-
     }
 }
